Extract SimpleAsm from a6502 into d6502
Extract the code for parsing and assembling a single instruction from
a6502 into a new module SimpleAsm and add it to the d6502 library.

The new module exposes an interface to assemble an instruction and to
parse an expression using the operand parser. It is configured with an
error reporting callback and a callback for label resolution.