[apple2tc] first version of disassembler
For now it is a simple jump tracing disassembler. It has the ability to
produce an assembler listing and a *simple C*" code.

The *simple C* code is a straight implementation of all 6502
instructions in C, stringed one after another. Registers and RAM are
global variables. The only purpose of this code generation mode is to
validate the concept, the disassembly and to eventually ensure that we
have discovered all basic blocks.

The code is split into basic blocks with a giant switch statement by
starting address. Thus it can only handle code for which we have
complete knowledge of all branch targets.

Self modifying code is detected statically and handled at the
instruction level.

As of this commit, the first 100,000 basic blocks execute identically in
*simple C* mode and in the emulator. After that soon we encounter an
indirect branch to an undetected code address. That was expected - we
need dynamic information from the emulator.

The generated C code is supported by a couple of extra files:
- `system.h`: a header file defining the interface of the generated
code.
- `system-inc.h`: a file included by the generated code, providing the
 supporting boilerplate.

For now these files are placed under `decoded/bolo` where we are running
experiments for decompiling the well known classic game Bolo:
https://en.wikipedia.org/wiki/Bolo_(1982_video_game)