[decoded/rom] WIP decompilation of the AppleII+ rom
It works well enough, even typing text, until Return is pressed at which
point it attempts to execute CHRGET from the zero page. Yes, CHRGET
is an entire routine residing in the zero page, only so it can
self-modify the address from which it is reading. TXTPTR aliases the
operand of a LDA in zero page.