[decoded/rom] updated with CHRGET
CHRGET is now successfully decompiled. Pressing Return even works for
some commands like HOME.

A new problem has appeared: address $DB59 uses an overlapping
instruction trick to skip the next instruction if executed. The next
instruction is 2 bytes and is contained entirely in the space of the
operand.

We need to detect this case and support it. Seems doable. If the next
instruction fits in the operand, special case the C code emission.