[apple2tc] ignore non-simple misaligned labels
So far non-simple misaligned labels are an indicator that we have
entered non-code, so handling them only makes matters worse. For now we
just detect them and ignore them. If there is a reason we might decide
to handle them again in the future.

BTW, this shows that we have a problem with invalid code. If we
encounter an invalid instruction or otherwise detect that the code is
bad, we should probably reject the entire rage since the last label.