a2emu: option to start tracing from reset
The `--rom` option enables tracing from reset, which allows us to record
the BASIC initialization, which is relied on by some games. If a binary
is also specified, pressing F3 generates the appropriate `CALL XXXX`
command in BASIC, so it can also be traced.

Cleanup a lot of accumulated hacks in a2emu and DebugState6502.