decapplib: debug/trace improvement
By default output the debugging information in a new mode - cycles +
address, without registers, unless the "--compat" option is specified,
which enables compatibility with the emulator traces.

The old style emulator traces are just not useful when debugging the
higher levels of C code.

Make sure when tracing the cycles are exactly reproducible.